Renamve directory to DateStoreProvider
