+++ src/components/BalanceSheet/BalanceSheet.tsx
@@ -5,7 +5,7 @@ import { type View as ViewType } from '@internal-types/general'
 import { useBalanceSheet } from '@hooks/api/businesses/[business-id]/reports/balance-sheet/useBalanceSheet'
 import { useReportsCompactHeader } from '@hooks/features/reports/useReportsCompactHeader'
 import { useResolvedReportView } from '@hooks/features/reports/useResolvedReportView'
-import { type DateSelectionMode, useGlobalDate } from '@providers/DateStore/GlobalDateStoreProvider'
+import { type DateSelectionMode, useGlobalDate } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
 import { TableProvider } from '@contexts/TableContext/TableContext'
 import { HStack, Stack } from '@ui/Stack/Stack'
 import { BALANCE_SHEET_ROWS_CONFIG } from '@components/BalanceSheet/constants'
