+++ src/components/MileageSummaryCard/MileageSummaryCard.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next'
 import { DateFormat } from '@utils/i18n/date/patterns'
 import { useMileageSummary } from '@hooks/api/businesses/[business-id]/mileage/summary/useMileageSummary'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
-import { useGlobalDate } from '@providers/DateStore/GlobalDateStoreProvider'
+import { useGlobalDate } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
 import ArrowRightCircleSubtle from '@icons/ArrowRightCircleSubtle'
 import { Button } from '@ui/Button/Button'
 import { HStack, VStack } from '@ui/Stack/Stack'
