+++ src/components/Tasks/TasksPending.tsx
@@ -6,7 +6,7 @@ import { getCompletedTasks, getIncompleteTasks } from '@utils/bookkeeping/tasks/
 import { DateFormat } from '@utils/i18n/date/patterns'
 import { useActiveBookkeepingPeriod } from '@hooks/features/bookkeeping/useActiveBookkeepingPeriod'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
-import { useGlobalDate } from '@providers/DateStore/GlobalDateStoreProvider'
+import { useGlobalDate } from '@providers/DateStoreProvider/GlobalDateStoreProvider'
 import { Heading } from '@ui/Typography/Heading'
 import { P } from '@ui/Typography/Text'
 import { BookkeepingStatus } from '@components/BookkeepingStatus/BookkeepingStatus'
